pyproject.toml
[project]
name = "song-player"
version = "0.1.0"
description = "Tunify - Music Player with Lyrics Sync"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-cloud-storage>=3.8.0",
    "google-genai>=1.60.0",
    "httpx>=0.28.0",
    "pymongo[srv]==3.12",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
